+++ infra/growth/test/apollo-adapter.test.js
@@ -1,4 +1,3 @@
-const BigNumber = require('bignumber.js')
 const chai = require('chai')
 const expect = chai.expect
 
@@ -7,219 +6,28 @@ const { CampaignRules } = require('../src/resources/rules')
 const { ApolloAdapter, campaignToApolloObject } = require('../src/apollo/adapter')
 const enums = require('../src/enums')
 const db = require('../src/models')
+const tokenNaturalUnits = require('../src/util/token')
 
-
-function tokenNaturalUnits(x) {
-  return BigNumber(x)
-    .times(BigNumber(10).pow(18))
-    .toFixed()
-}
 
 describe('Apollo adapter', () => {
-
 
   describe('April campaign rules', () => {
 
     before(async () => {
-      const config = {
-        numLevels: 3,
-        levels: {
-          0: {
-            rules: [
-              {
-                id: 'ProfilePublished',
-                class: 'SingleEvent',
-                config: {
-                  eventType: 'ProfilePublished',
-                  reward: null,
-                  nextLevelCondition: false,
-                  visible: true,
-                  scope: 'user'
-                }
-              },
-              {
-                id: 'EmailAttestation',
-                class: 'SingleEvent',
-                config: {
-                  eventType: 'EmailAttestationPublished',
-                  reward: null,
-                  visible: true,
-                  limit: 1,
-                  nextLevelCondition: false,
-                  scope: 'user'
-                }
-              },
-              {
-                id: 'BasicProfile',
-                class: 'MultiEvents',
-                config: {
-                  eventTypes: ['ProfilePublished', 'EmailAttestationPublished'],
-                  visible: false,
-                  repeatable: false,
-                  numEventsRequired: 2,
-                  reward: null,
-                  nextLevelCondition: true,
-                  scope: 'user',
-                  unlockConditionMsg: [
-                    {
-                      conditionTranslateKey: 'growth.profile.and.email.requirement',
-                      conditionIcon: 'images/growth/email-icon-small.svg'
-                    }
-                  ],
-                }
-              }
-            ],
-          },
-          1: {
-            rules: [
-              {
-                id: 'PhoneAttestation',
-                class: 'SingleEvent',
-                config: {
-                  eventType: 'PhoneAttestationPublished',
-                  reward: {
-                    amount: tokenNaturalUnits(25),
-                    currency: 'OGN'
-                  },
-                  limit: 1,
-                  nextLevelCondition: false,
-                  visible: true,
-                  scope: 'user'
-                }
-              },
-              {
-                id: 'FacebookAttestation',
-                class: 'SingleEvent',
-                config: {
-                  eventType: 'FacebookAttestationPublished',
-                  reward: {
-                    amount: tokenNaturalUnits(25),
-                    currency: 'OGN'
-                  },
-                  limit: 1,
-                  nextLevelCondition: false,
-                  visible: true,
-                  scope: 'user'
-                }
-              },
-              {
-                id: 'AirbnbAttestation',
-                class: 'SingleEvent',
-                config: {
-                  eventType: 'AirbnbAttestationPublished',
-                  reward: {
-                    amount: tokenNaturalUnits(25),
-                    currency: 'OGN'
-                  },
-                  limit: 1,
-                  nextLevelCondition: false,
-                  visible: true,
-                  scope: 'user'
-                }
-              },
-              {
-                id: 'TwitterAttestation',
-                class: 'SingleEvent',
-                config: {
-                  eventType: 'TwitterAttestationPublished',
-                  reward: {
-                    amount: tokenNaturalUnits(25),
-                    currency: 'OGN'
-                  },
-                  limit: 1,
-                  nextLevelCondition: false,
-                  visible: true,
-                  scope: 'user'
-                }
-              },
-              {
-                id: 'TwoAttestations',
-                class: 'MultiEvents',
-                config: {
-                  eventTypes: [
-                    'PhoneAttestationPublished',
-                    'FacebookAttestationPublished',
-                    'AirbnbAttestationPublished',
-                    'TwitterAttestationPublished'
-                  ],
-                  numEventsRequired: 2,
-                  reward: null,
-                  nextLevelCondition: true,
-                  unlockConditionMsg: [
-                    {
-                      conditionTranslateKey: 'growth.two.attestations.requirement',
-                      conditionIcon: 'images/growth/attestation-icon.svg'
-                    }
-                  ],
-                  visible: false,
-                  scope: 'user'
-                }
-              }
-            ],
-          },
-          2: {
-            rules: [
-              {
-                id: 'Referral',
-                class: 'Referral',
-                config: {
-                  levelRequired: 1,
-                  reward: {
-                    amount: tokenNaturalUnits(50),
-                    currency: 'OGN'
-                  },
-                  limit: 25,
-                  nextLevelCondition: false,
-                  visible: true,
-                  scope: 'campaign'
-                }
-              },
-              {
-                id: 'ListingPurchase',
-                class: 'SingleEvent',
-                config: {
-                  eventType: 'ListingPurchased',
-                  reward: {
-                    amount: tokenNaturalUnits(100),
-                    currency: 'OGN'
-                  },
-                  limit: 1,
-                  nextLevelCondition: false,
-                  visible: true,
-                  scope: 'campaign'
-                }
-              },
-              {
-                id: 'ListingSold',
-                class: 'SingleEvent',
-                config: {
-                  eventType: 'ListingSold',
-                  reward: {
-                    amount: tokenNaturalUnits(100),
-                    currency: 'OGN'
-                  },
-                  limit: 1,
-                  nextLevelCondition: false,
-                  visible: true,
-                  scope: 'campaign'
-                }
-              }
-            ],
-          }
-        }
-      }
+      const aprilConfig = require('../campaigns/april.js')
+
       this.campaignStart = new Date()
       this.campaignEnd = new Date(this.campaignStart.getTime()+100000)
       this.duringCampaign = new Date(this.campaignStart.getTime()+100)
       this.beforeCampaign = new Date(this.campaignStart.getTime()-100000)
       const row = {
         id: 1,
-        rules: JSON.stringify(config),
+        rules: JSON.stringify(aprilConfig),
         startDate: this.campaignStart,
         endDate: this.campaignEnd,
         currency: 'OGN'
       }
-      this.crules = new CampaignRules(row, config)
+      this.crules = new CampaignRules(row, aprilConfig)
       expect(this.crules).to.be.an('object')
       expect(this.crules.numLevels).to.equal(3)
       expect(this.crules.levels[0]).to.be.an('object')
