+++ infra/growth/test/rules-may.test.js
@@ -1,285 +1,16 @@
-const BigNumber = require('bignumber.js')
 const chai = require('chai')
 const expect = chai.expect
 
 const { GrowthEventTypes, GrowthEventStatuses } = require('../src/enums')
 const { CampaignRules } = require('../src/resources/rules')
+const tokenNaturalUnits = require('../src/util/token')
 
-function tokenNaturalUnits(x) {
-  return BigNumber(x)
-    .times(BigNumber(10).pow(18))
-    .toFixed()
-}
 
 describe('May campaign rules', () => {
 
   before(() => {
-    const config = {
-        numLevels: 3,
-        levels: {
-          0: {
-            rules: [
-              {
-                id: 'ProfilePublished',
-                class: 'SingleEvent',
-                config: {
-                  eventType: 'ProfilePublished',
-                  reward: null,
-                  nextLevelCondition: false,
-                  visible: true,
-                  scope: 'campaign'
-                }
-              },
-              {
-                id: 'EmailAttestation',
-                class: 'SingleEvent',
-                config: {
-                  eventType: 'EmailAttestationPublished',
-                  reward: null,
-                  visible: true,
-                  nextLevelCondition: false,
-                  scope: 'campaign'
-                }
-              },
-              {
-                id: 'BothRules',
-                class: 'MultiEvents',
-                config: {
-                  eventTypes: ['ProfilePublished', 'EmailAttestationPublished'],
-                  visible: false,
-                  numEventsRequired: 2,
-                  reward: null,
-                  nextLevelCondition: true,
-                  scope: 'user',
-                  unlockConditionMsg: [
-                    {
-                      conditionTranslateKey: 'growth.profile.requirement',
-                      conditionIcon: 'images/growth/email-icon-small.svg'
-                    },
-                    {
-                      conditionTranslateKey: 'growth.email.requirement',
-                      conditionIcon: 'images/growth/email-icon-small.svg'
-                    }
-                  ]
-                }
-              }
-            ]
-          },
-          1: {
-            rules: [
-              {
-                id: 'PhoneAttestation',
-                class: 'SingleEvent',
-                config: {
-                  eventType: 'PhoneAttestationPublished',
-                  reward: {
-                    amount: tokenNaturalUnits(25),
-                    currency: 'OGN'
-                  },
-                  visible: true,
-                  limit: 1,
-                  nextLevelCondition: false,
-                  scope: 'campaign'
-                }
-              },
-              {
-                id: 'FacebookAttestation',
-                class: 'SingleEvent',
-                config: {
-                  eventType: 'FacebookAttestationPublished',
-                  reward: {
-                    amount: tokenNaturalUnits(25),
-                    currency: 'OGN'
-                  },
-                  visible: true,
-                  limit: 1,
-                  nextLevelCondition: false,
-                  scope: 'campaign'
-                }
-              },
-              {
-                id: 'AirbnbAttestation',
-                class: 'SingleEvent',
-                config: {
-                  eventType: 'AirbnbAttestationPublished',
-                  reward: {
-                    amount: tokenNaturalUnits(25),
-                    currency: 'OGN'
-                  },
-                  visible: true,
-                  limit: 1,
-                  nextLevelCondition: false,
-                  scope: 'campaign'
-                }
-              },
-              {
-                id: 'TwitterAttestation',
-                class: 'SingleEvent',
-                config: {
-                  eventType: 'TwitterAttestationPublished',
-                  reward: {
-                    amount: tokenNaturalUnits(25),
-                    currency: 'OGN'
-                  },
-                  visible: true,
-                  limit: 1,
-                  nextLevelCondition: false,
-                  scope: 'campaign'
-                }
-              },
-              {
-                id: 'GoogleAttestation',
-                class: 'SingleEvent',
-                config: {
-                  eventType: 'GoogleAttestationPublished',
-                  reward: {
-                    amount: tokenNaturalUnits(25),
-                    currency: 'OGN'
-                  },
-                  visible: true,
-                  limit: 1,
-                  nextLevelCondition: false,
-                  scope: 'campaign'
-                }
-              },
-              {
-                id: 'TwoAttestations',
-                class: 'MultiEvents',
-                config: {
-                  eventTypes: [
-                    'PhoneAttestationPublished',
-                    'FacebookAttestationPublished',
-                    'AirbnbAttestationPublished',
-                    'TwitterAttestationPublished',
-                    'GoogleAttestationPublished'
-                  ],
-                  visible: false,
-                  numEventsRequired: 2,
-                  reward: null,
-                  nextLevelCondition: true,
-                  scope: 'user',
-                  unlockConditionMsg: [
-                    {
-                      conditionTranslateKey: 'growth.two.attestations.requirement',
-                      conditionIcon: 'images/growth/attestation-icon.svg'
-                    }
-                  ]
-                }
-              }
-            ]
-          },
-          2: {
-            rules: [
-              {
-                id: 'Referral',
-                class: 'Referral',
-                config: {
-                  levelRequired: 2,
-                  reward: {
-                    amount: tokenNaturalUnits(50),
-                    currency: 'OGN'
-                  },
-                  limit: 25,
-                  visible: true,
-                  nextLevelCondition: false,
-                  scope: 'campaign'
-                }
-              },
-              {
-                id: 'ListingPurchaseTShirt',
-                class: 'ListingPurchase',
-                config: {
-                  eventType: 'ListingPurchased',
-                  listingId: '1-000-1',
-                  reward: {
-                    amount: tokenNaturalUnits(25),
-                    currency: 'OGN'
-                  },
-                  visible: true,
-                  limit: 100,
-                  nextLevelCondition: false,
-                  scope: 'campaign'
-                }
-              },
-              {
-                id: 'ListingPurchaseGC',
-                class: 'ListingPurchase',
-                config: {
-                  listingId: '1-000-2',
-                  reward: {
-                    amount: tokenNaturalUnits(50),
-                    currency: 'OGN'
-                  },
-                  visible: true,
-                  limit: 100,
-                  nextLevelCondition: false,
-                  scope: 'campaign'
-                }
-              },
-              {
-                id: 'ListingPurchaseDonation',
-                class: 'ListingPurchase',
-                config: {
-                  listingId: '1-000-3',
-                  reward: {
-                    amount: tokenNaturalUnits(500),
-                    currency: 'OGN'
-                  },
-                  visible: true,
-                  limit: 100,
-                  nextLevelCondition: false,
-                  scope: 'campaign'
-                }
-              },
-              {
-                id: 'ListingPurchaseHousing',
-                class: 'ListingPurchase',
-                config: {
-                  listingId: '1-000-4',
-                  reward: {
-                    amount: tokenNaturalUnits(100),
-                    currency: 'OGN'
-                  },
-                  visible: true,
-                  limit: 100,
-                  nextLevelCondition: false,
-                  scope: 'campaign'
-                }
-              },
-              {
-                id: 'ListingPurchaseInfluencer',
-                class: 'ListingPurchase',
-                config: {
-                  listingId: '1-000-5',
-                  reward: {
-                    amount: tokenNaturalUnits(100),
-                    currency: 'OGN'
-                  },
-                  visible: true,
-                  limit: 100,
-                  nextLevelCondition: false,
-                  scope: 'campaign'
-                }
-              },
-              {
-                id: 'ListingPurchaseArt',
-                class: 'ListingPurchase',
-                config: {
-                  listingId: '1-000-6',
-                  reward: {
-                    amount: tokenNaturalUnits(1000),
-                    currency: 'OGN'
-                  },
-                  visible: true,
-                  limit: 100,
-                  nextLevelCondition: false,
-                  scope: 'campaign'
-                }
-              }
-            ]
-          }
-        }
-      }
+    const mayConfig = require('../campaigns/may.js')
+
     this.campaignStart = new Date()
     this.campaignEnd = new Date(this.campaignStart.getTime()+100000)
     this.duringCampaign = new Date(this.campaignStart.getTime()+100)
@@ -288,12 +19,12 @@ describe('May campaign rules', () => {
 
     const row = {
       id: 1,
-      rules: JSON.stringify(config),
+      rules: JSON.stringify(mayConfig),
       startDate: this.campaignStart,
       endDate: this.campaignEnd,
       currency: 'OGN'
     }
-    this.crules = new CampaignRules(row, config)
+    this.crules = new CampaignRules(row, mayConfig)
     expect(this.crules).to.be.an('object')
     expect(this.crules.numLevels).to.equal(3)
     expect(this.crules.levels[0]).to.be.an('object')
